extract IdentifierSystemInput component
